+++ lib/src/core/util/collections.ts
@@ -1,4 +1,4 @@
-import { EntityCollection, PropertiesOrBuilders, PropertyOrBuilder } from "../../types";
+import {DefaultSelectedViewBuilder, DefaultSelectedViewParams, EntityCollection, PropertiesOrBuilders, PropertyOrBuilder } from "../../types";
 import { mergeDeep } from "./objects";
 import { isPropertyBuilder } from "./entities";
 
@@ -83,5 +83,18 @@ export function sortProperties<M extends Record<string, any>>(properties: Proper
     } catch (e) {
         console.error("Error sorting properties", e);
         return properties;
+    }
+}
+
+export function resolveDefaultSelectedView(
+    defaultSelectedView: string | DefaultSelectedViewBuilder | undefined,
+    params: DefaultSelectedViewParams
+) {
+    if (!defaultSelectedView) {
+        return undefined;
+    } else if (typeof defaultSelectedView === "string") {
+        return defaultSelectedView;
+    } else {
+        return defaultSelectedView(params);
     }
 }