+++ lib/src/firebase_app/FirebaseCMSAppProps.tsx
@@ -17,6 +17,7 @@ import { FirestoreTextSearchController } from "./types/text_search";
 import { Authenticator, FirebaseSignInOption, FirebaseSignInProvider } from "./types/auth";
 import { FirebaseLoginViewProps } from "./components/FirebaseLoginView";
 import { useNavigationContext } from "../hooks";
+import { FirebaseApp } from "firebase/app";
 
 /**
  * Main entry point that defines the CMS configuration
@@ -97,7 +98,7 @@ export type FirebaseCMSAppProps = {
      * using the local emulator or retrieving the used configuration.
      * @param config
      */
-    onFirebaseInit?: (config: object) => void;
+    onFirebaseInit?: (config: object, app: FirebaseApp) => void;
 
     /**
      * Use this to enable Firebase App Check